requirements.txt
requests==2.24.0
PyYAML==5.3.1
invoke==1.4.1
pytest==6.1.2
